feat: カードパスを再構成（閲覧=/card/[cardId]、編集=/card/[cardId]/edit）
- /card/[cardId] → 閲覧画面（CardViewClient）+ OGP generateMetadata
- /card/[cardId]/edit → 編集画面（CardEditorClient）
- /card/[cardId]/view/ ディレクトリを削除
- 全リンク・リダイレクト・テストを新パスに更新
