+++ src/app/card/[cardId]/edit/page.tsx
@@ -1,0 +1,51 @@
+import { createClient } from '@/lib/supabase/server'
+import { notFound, redirect } from 'next/navigation'
+import { v1Template } from '@/templates/v1'
+import { v2Template } from '@/templates/v2'
+import CardEditorClient from './CardEditorClient'
+
+const templateMap = {
+  v1: v1Template,
+  v2: v2Template,
+}
+
+export default async function CardPage({ params }: { params: Promise<{ cardId: string }> }) {
+  const { cardId } = await params
+  const supabase = await createClient()
+
+  const { data: { user } } = await supabase.auth.getUser()
+
+  const { data: card, error: cardError } = await supabase
+    .from('cards')
+    .select('*')
+    .eq('id', cardId)
+    .single()
+
+  if (cardError) console.error('[CardPage] card fetch error:', cardError)
+  if (!card) notFound()
+
+  // 非公開カードは本人のみ
+  if (card.visibility === 'private' && card.user_id !== user?.id) {
+    redirect('/auth/login')
+  }
+
+  const template = templateMap[card.template_id as keyof typeof templateMap]
+  if (!template) notFound()
+
+  const isOwner = user?.id === card.user_id
+
+  const { data: announcements } = await supabase
+    .from('announcements')
+    .select('id, title, body, published_at')
+    .eq('is_active', true)
+    .order('published_at', { ascending: false })
+
+  return (
+    <CardEditorClient
+      card={card}
+      template={template}
+      isOwner={isOwner}
+      announcements={announcements ?? []}
+    />
+  )
+}
